requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.4.0
python-decouple==3.8
coverage==7.3.2
factory-boy==3.3.0